+++ wavemon.h
@@ -255,7 +255,7 @@ static inline void waddstr_b(WINDOW * win, const char *s)
 }
 
 extern void waddstr_center(WINDOW * win, int y, const char *s);
-extern const char *curtail(const char *str, const char *sep, int len);
+extern const char *curtail(const char *str, const char *sep, size_t len);
 
 extern void waddbar(WINDOW *win, int y, float v, float min, float max,
 		    int8_t *cscale, bool rev);
